+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## v7.15.0 Next release
+
+- Adds `all_children` function to the User class for retrieving paginated lists of children users, and deprecate the beta function
+- Adds `get_next_page_of_children` function to the User class for retrieving next paginated lists of children users, and deprecate the beta function
+
 ## v7.14.1 (2023-10-30)
 
 - Fixes a bug where `get_next_page` functions threw an error, preventing users from retrieving the final page of results